requirements.txt
PySide6
pyinstaller
psycopg2
pillow